\name{Europe_Forest}
\alias{Europe_Forest}
\title{Normalized ecosystem function data for Europe forest}

\description{
This is the dataset of \code{Europe_Forest_raw} dataset after ecosystem functions being normolized, and there are two ecosystem functions "soil_cn_ff_10" and "wue" being normalized negatively.
}

\usage{
data("Europe_Forest")
}

\format{
a data.frame with 209 plots (rows) and 30 normalized ecosystem functions, properties or corresponding informations (columns).
}


\references{

Baeten, L., Verheyen, K., Wirth, C., Bruelheide, H., Bussotti, F., Finér, L.,…, Scherer-Lorenzen, M. (2013). A novel comparative research platform designed to determine the functional significance of tree species diversity in European forests. Perspectives in Plant Ecology, Evolution and Systematics, 15, 281–291. \cr

van der Plas, F., Manning, P., Soliveres, S., Allan, E., Scherer-Lorenzen, M., Verheyen, K., . . . Fischer, M. (2016). Biotic homogenization can decrease landscape-scale forest multifunctionality. Proceedings of the National Academy of Sciences of the United States of America, 113, E3557–E3562. \cr

Ratcliffe, S. Wirth, C., Jucker, T. van der Plas, F., Scherer-Lorenzen, M. Verheyen, K. et al. (2017). Biodiversity and ecosystem functioning relations in European forests depend on environmental context. Ecology Letters, 20, 1414–1426. \cr

Chao, A., Chiu, C. H., Hu, K. H., van der Plas, F., Cadotte, M. W., Mitesser, O., et al. (2023). Hill-Chao numbers in multifunctionality allows decomposing gamma multifunctionality into alpha and beta components. Under revision, Ecology Letters.

}

\keyword{datasets}

