\name{forest_function_data_normalized}
\alias{forest_function_data_normalized}
\title{Normalized ecosystem function data for six European forests}

\description{
This is the normalized version of the raw dataset (\code{forest_function_data_raw}) when all ecosystem functions are normalized to the range of [0, 1].
}

\usage{
data("forest_function_data_normalized")
}

\format{
a data.frame with 209 plots (rows) and 32 columns, in addition to row and column names; the first 5 columns show the relevant plot information, followed by 26 normalized ecosystem functions. And the last column shows the corresponding country for each plot.
}


\references{

Ratcliffe, S. Wirth, C., Jucker, T. van der Plas, F., Scherer-Lorenzen, M. Verheyen, K. et al. (2017). Biodiversity and ecosystem functioning relations in European forests depend on environmental context. Ecology Letters, 20, 1414–1426. \cr

Scherer-Lorenzen, M. et al. (2023). The functional significance of tree species diversity in European forests - the FunDivEUROPE dataset [Dataset]. Dryad. https://doi.org/10.5061/dryad.9ghx3ffpz

}

\keyword{datasets}

