% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfun.R
\name{func_normalized}
\alias{func_normalized}
\title{For functions normalized}
\usage{
func_normalized(
  data,
  fun_cols = 1:ncol(data),
  negative = NULL,
  by_group = NULL
)
}
\arguments{
\item{data}{data can be input as a matrix/data.frame (assemblages by functions). Data should be without NA, if does, the missing values need to impute first and then use the function.}

\item{fun_cols}{the order number of the columns which be used as the ecosystem function.}

\item{negative}{names of columns which need to be normalized negatively.}

\item{by_group}{name of the column to normalized by group. Default is \code{NULL}.}
}
\value{
a data.frame with normalized functions.
}
\description{
\code{func_normalized}:\cr
To get normalized functions between 0 and 1.
}
\examples{
\dontrun{
data("Europe_Forest_raw")
GER_ITA_Forest_raw <- filter(Europe_Forest_raw,Country=="GER"|Country=="ITA")
func_normalized(data = GER_ITA_Forest_raw, fun_cols = 4:29,
                negative = c("soil_cn_ff_10","wue"), by_group = "Country")

}
}
