% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfun.plot.R
\name{ggMF}
\alias{ggMF}
\title{ggplot2 extension for a MF object}
\usage{
ggMF(
  output,
  by_group = NULL,
  facets_scale = "fixed",
  fit = "LMM.intercept",
  text = "Slope"
)
}
\arguments{
\item{output}{the output from \code{MF_single} or \code{MF_multiple}.}

\item{by_group}{whether the MF object to classify by group or not. If does, input name of the column used for grouping. Default is \code{NULL}.
For \code{fit} is selected to be linear mixed model, you must offer the \code{by_group} argument.}

\item{facets_scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{fit}{method of the fitted line. Select \code{fit = "lm"} for the linear model, or \code{fit = "LMM.intercept"}, \code{fit = "LMM.slope"} and \code{fit = "LMM.both"} for the linear mixed model with random effect 'intercept', 'slope' and 'both intercept and slope', respectively.
Default is \code{fit = "LMM.intercept"}.}

\item{text}{type of text information show in the plots. Select \code{text = "Slope"} to see estimated of slopes, or \code{text = "R.squared"} to see model performance. Default is \code{text = "Slope"}.}
}
\value{
For \code{MF_single} output, return a figure for multi-functionality with functions that are uncorrected and corrected for correlations.
For \code{MF_multiple} output, return a list of figures with uncorrected and corrected for correlations.
}
\description{
\code{ggMF}:the \code{\link[ggplot2]{ggplot}} extension for \code{MF} object to plot the correlation between species diversity and multifunctionality.
}
\examples{

\dontrun{

## single ecosystem
data("Europe_Forest")
data("Europe_Forest_species")
GER_ITA_Forest_function <- filter(Europe_Forest,Country=="GER"|Country=="ITA")[,4:29]
GER_ITA_Forest_species<-Europe_Forest_species[49:229,]
output1<-MF_single(func_data = GER_ITA_Forest_function, species_data = GER_ITA_Forest_species)


## Display fitted line of linear model

GER_ITA_Forest<-filter(Europe_Forest,Country=="GER"|Country=="ITA")
output1 <- data.frame(output1,
 Country = rep(GER_ITA_Forest$Country, each = 6))
ggMF(output1, facets_scale = 'fixed', fit = "lm",by_group="Country")


## multiple ecosystems
data("Europe_Forest")
data("Europe_Forest_species")
GER_ITA_Forest <- filter(Europe_Forest,Country=="GER"|Country=="ITA")
GER_ITA_Forest_species<-Europe_Forest_species[49:229,]
output2 <- MF_multiple(func_data = GER_ITA_Forest[,4:30],
                          species_data = GER_ITA_Forest_species,
                          by_group = "Country")

## Display fitted line of linear mixed model with random effect 'intercept'
ggMF(
output = output2,
by_group = "Country",
facets_scale = "free_x",
fit = "LMM.intercept",
text = "Slope"
)

}

}
