% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfun.R
\name{MF_single}
\alias{MF_single}
\title{MF measures (single)}
\usage{
MF_single(func_data, species_data = NULL, q = c(0, 1, 2))
}
\arguments{
\item{func_data}{the ecosystem function data can be input as a vector of functions (for a single assemblage), matrix/data.frame (assemblages by functions).\cr
The data input must be normalized between 0 and 1 already and must contain only the ecosystem function columns.\cr
For \code{species_data} is not \code{NULL}, the rownames of func_data should be names of plotID.}

\item{species_data}{the species abundance data must include three columns: 'plotID', 'species' and 'abundance'. Default is \code{NULL}.}

\item{q}{a numerical vector specifying the diversity orders. Default is 0, 1 and 2.}
}
\value{
a data.frame with columns 'Type', 'Order.q' and 'qMF'. For \code{species_data} is not \code{NULL}, the data.frame will contain 'Species.diversity'.
}
\description{
\code{MF_single}:\cr
Multi-functionality measures in a single ecosystem. This is a function in the framework of Hill-Chao numbers
}
\examples{

data("Europe_Forest")
data("Europe_Forest_species")
Europe_Forest_function <- Europe_Forest[,4:29]
rownames(Europe_Forest_function) <- Europe_Forest$plotid
MF_single(func_data = Europe_Forest_function, species_data = Europe_Forest_species)

}
