% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfun.plot.R
\name{MFggplot}
\alias{MFggplot}
\title{ggplot2 extension for a MF1_single or MF2_multiple object}
\usage{
MFggplot(
  output,
  model = "LMM.both",
  caption = "slope",
  by_group = NULL
)
}
\arguments{
\item{output}{the output obtained from \code{MF1_single} or \code{MF2_multiple}. \cr 
If using output obtained from \code{MF1_single} and want to draw the plot with \code{by_group}, must add the \code{by_group} column in output.}

\item{model}{specifying the fitting model. Specify \code{model = "lm"} for the linear model, or \code{"LMM.intercept"},
\code{"LMM.slope"}, \code{"LMM.both"} for the linear mixed model with random effects for intercept, slope, and both, respectively. Default is \code{model = "LMM.both"}.}

\item{caption}{caption information that will be shown in the BEF plots. Select \code{caption = "slope"} to show the estimated slopes, or \code{caption = "R.squared"} to show the marginal and conditional R-squared. Default is \code{caption = "slope"}.}

\item{by_group}{name of the column for normalization and decomposition i.e., function normalization and multifunctionality decomposition will be performed within each group classified by the categories of that column. Default is \code{NULL}. \cr
For linear mixed model is selected in the \code{model}, the \code{by_group} argument must be specified.}
}
\value{
For \code{MF1_single} object, return a figure that shows multifunctionality for uncorrected and corrected for correlations two cases. And also contains the fitted lines for the chosen fitting model on the figure. \cr
For \code{MF2_multiple} object, return a list of figures which show alpha, gamma, beta multifunctionality for uncorrected and corrected for correlations two cases, respectively. And also contains the fitted lines for the chosen fitting model on the figures.
}
\description{
\code{MFggplot}:the \code{\link[ggplot2]{ggplot}} extension for \code{MF1_single} or \code{MF2_multiple} object to plot the BEF relationship between biodiversity and multifunctionality.
}
\examples{

\dontrun{

## single ecosystem
data("forest_function_data_normalized")
data("forest_biodiversity_data")
GER_ITA_forest_function_normalized <- filter(forest_function_data_normalized, 
                                             country=="GER"|country=="ITA")
GER_ITA_forest_biodiversity <- forest_biodiversity_data[49:229,]
output1 <- MF1_single(func_data = GER_ITA_forest_function_normalized[,6:31], 
                      species_data = GER_ITA_forest_biodiversity)

## Display fitted line of linear model
output1 <- data.frame(output1, country=rep(GER_ITA_forest_function_normalized$country, each = 6))
MFggplot(output1, model = "lm", by_group="country")


## multiple ecosystems
data("forest_function_data_normalized")
data("forest_biodiversity_data")
GER_ITA_forest_function_normalized <- filter(forest_function_data_normalized, 
                                             country=="GER"|country=="ITA")
GER_ITA_forest_biodiversity <- forest_biodiversity_data[49:229,]
output2 <- MF2_multiple(func_data = GER_ITA_forest_function_normalized[,6:32],
                        species_data = GER_ITA_forest_biodiversity,
                        by_group = "country")

## Display fitted line of linear mixed model with random effect 'both'
MFggplot(output2, model = "LMM.both", by_group = "country")


}

}
