\name{forest_function_data_normalized}
\alias{forest_function_data_normalized}
\title{Normalized ecosystem function data for six European forests}

\description{
This dataset represents the normalized version of the raw dataset (\code{forest_function_data_raw}). All raw ecosystem functions are normalized to the range of [0, 1], whereas other variables remain unchanged.
}

\usage{
data("forest_function_data_normalized")
}

\format{
a data.frame with 209 plots (rows) and 32 columns, in addition to row and column names; the first 5 columns show the relevant plot information, followed by 26 normalized ecosystem functions. The last column shows the corresponding country for each plot.
}



\keyword{datasets}

