% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfun.plot.R
\name{ggMF}
\alias{ggMF}
\title{ggplot2 extension for a MF object}
\usage{
ggMF(
  output,
  by_group = NULL,
  facets_scale = "fixed",
  fit = "LMM.intercept",
  text = "Slope"
)
}
\arguments{
\item{output}{the output from \code{MF_single} or \code{MF_multiple}. Note that, the column \code{Species.diversity} or the three columns \code{Species_gamma}, \code{Species_alpha} and \code{Species_beta} must be included in it, respectively.}

\item{by_group}{name of the column to be plot in difference colors by group. Default is \code{NULL}.\cr
For \code{fit} is selected to be linear mixed model, you must offer the \code{by_group} argument.}

\item{facets_scale}{Are scales shared across all facets (the default, \code{"fixed"}), or do they vary across rows (\code{"free_x"}), columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{fit}{method of the fitted line. Select \code{fit = "lm"} for the linear model, or \code{fit = "LMM.intercept"}, \code{fit = "LMM.slope"} and \code{fit = "LMM.both"} for the linear mixed model with random effect 'intercept', 'slope' and 'both intercept and slope', respectively. Default is \code{fit = "LMM.intercept"}}

\item{text}{what text information would show in plots? Select \code{text = "Slope"} to see estimated of slopes, or \code{text = "R.squared"} to see model performance.Default is \code{text = "Slope"}.}
}
\value{
For \code{MF_single} output, return a figure for multi-functionality with uncorrelated and correlated functions. For \code{MF_multiple} output, return a list of uncorrelated and correlated figures.
}
\description{
\code{ggMF}: the \code{\link[ggplot2]{ggplot}} extension for \code{MF} object to plot the correlation between species diversity and multi-functionality
}
\examples{

## single ecosystem
data("Europe_Forest")
data("Europe_Forest_species")
Europe_Forest_function <- Europe_Forest[,4:29]
rownames(Europe_Forest_function) <- Europe_Forest$plotid
output1 <- MF_single(func_data = Europe_Forest_function, species_data = Europe_Forest_species)

## Display fitted line of linear mixed model with random effect 'intercept'.
ggMF(output1, facets_scale = 'fixed', fit = "LMM.intercept")


## multiple ecosystems
Europe_Forest_function$Country <- Europe_Forest$Country
output2 <- MF_multiple(func_data = Europe_Forest_function,
species_data = Europe_Forest_species, by_group = "Country")

## Display fitted line of linear model
ggMF(output2, by_group = "Country", facets_scale = 'fixed', fit = "LMM.intercept")


}
