% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mainfun.R
\name{MF1_single}
\alias{MF1_single}
\title{multifunctionality measures for a single ecosystem}
\usage{
MF1_single(func_data, species_data = NULL, weight = 1, q = c(0, 1, 2))
}
\arguments{
\item{func_data}{ecosystem function data can be input as a vector of functions (for a single ecosystem),
matrix/data.frame (ecosystems by functions for multiple ecosystems). All function values must be normalized between 0 and 1.\cr
For \code{species_data} is not \code{NULL}, the row names of \code{func_data} should be the names of plotID.}

\item{species_data}{the species abundance data must include three columns: 'plotID', 'species' and 'abundance'.
Default is \code{NULL}.}

\item{weight}{a numerical vector specifying weights for ecosystem functions.
The length of \code{weight} needs to be equal to the number of functions.
Default is \code{weight = 1}, which means equal weight for all ecosystem functions.}

\item{q}{a numerical vector specifying the diversity orders. Default is q = 0, 1 and 2.}
}
\value{
a data.frame with columns ’plotID’, ’Type’ (uncorrected or corrected for correlations),
’Order.q’ and ’qMF’. For \code{species_data} is not \code{NULL},
the data.frame will show an additional column contain ’Species.diversity’.
}
\description{
\code{MF1_single}:\cr
\code{MF1_single} computes multifunctionality measures for given function weights in a single ecosystem separately for two cases
(i) correlations between functions are not corrected for, and (ii) correlations between functions are corrected.
}
\examples{
\dontrun{
data("forest_function_data_normalized")
data("forest_biodiversity_data")
GER_ITA_forest_function_normalized <- filter(forest_function_data_normalized, 
                                             country=="GER"|country=="ITA")
GER_ITA_forest_biodiversity <- forest_biodiversity_data[49:229,]
MF1_single(func_data = GER_ITA_forest_function_normalized[,6:31], 
           species_data = GER_ITA_forest_biodiversity)
}
}
